+++ nix-integration-test/nixpkgs-commit.nix
@@ -1,0 +1,4 @@
+fetchTarball {
+ url = "https://github.com/typetetris/nixpkgs/tarball/685d2b8895faf86998877d23a7a51317e5b31425";
+ sha256 = "1yl49ldh0y5v76jx3fvqf76i21mypj869jj7ffbs8slpsbd5496k";
+}